fix: sync entities and added tolerance to hasCustomerProductEnded
