+++ .gitignore
@@ -86,7 +86,6 @@ next-env.d.ts
 shared/drizzle
 
 
-run.sh
 commands.sh
 env.sh
 server/env.sh
