+++ server/src/internal/billing/v2/actions/sync/compute/initImmediateSyncCustomerProduct.ts
@@ -35,7 +35,7 @@ export const initImmediateSyncCustomerProduct = ({
 	stripeSubscription: Stripe.Subscription;
 	currentEpochMs: number;
 }): FullCusProduct => {
-	const { plan, fullProduct, featureQuantities } = productContext;
+	const { plan, fullProduct, featureQuantities, entity } = productContext;
 
 	const trialEndsAt = getTrialEndsAtFromStripe({ stripeSubscription });
 	const { canceledAt, endedAt } = getCancelFieldsFromStripe({
@@ -49,6 +49,7 @@ export const initImmediateSyncCustomerProduct = ({
 			fullCustomer,
 			fullProduct,
 			featureQuantities,
+			entity,
 			resetCycleAnchor: resetCycleAnchorMs,
 			now: currentEpochMs,
 			freeTrial: null,
