+++ server/src/internal/billing/v2/utils/initFullCustomerProduct/initScheduledCustomerProduct.ts
@@ -1,10 +1,12 @@
 import {
 	BillingVersion,
 	CusProductStatus,
+	type Entity,
 	type FeatureOptions,
 	type FullCusProduct,
 	type FullCustomer,
 	type FullProduct,
+	truncateMsToSecondPrecision,
 } from "@autumn/shared";
 import type { AutumnContext } from "@/honoUtils/HonoEnv";
 import { initFullCustomerProduct } from "./initFullCustomerProduct";
@@ -21,6 +23,7 @@ export const initScheduledCustomerProduct = ({
 	fullCustomer,
 	fullProduct,
 	featureQuantities,
+	entity,
 	startsAt,
 	endsAt,
 	currentEpochMs,
@@ -33,6 +36,7 @@ export const initScheduledCustomerProduct = ({
 	fullCustomer: FullCustomer;
 	fullProduct: FullProduct;
 	featureQuantities: FeatureOptions[];
+	entity?: Entity;
 	startsAt: number;
 	endsAt: number | null | undefined;
 	currentEpochMs: number;
@@ -45,20 +49,27 @@ export const initScheduledCustomerProduct = ({
 	subscriptionId?: string;
 	subscriptionScheduleId?: string;
 }): FullCusProduct => {
+	const startsAtSecondsPrecision = truncateMsToSecondPrecision(startsAt);
+	const endsAtSecondsPrecision =
+		endsAt === null || endsAt === undefined
+			? undefined
+			: truncateMsToSecondPrecision(endsAt);
+
 	return initFullCustomerProduct({
 		ctx,
 		initContext: {
 			fullCustomer,
 			fullProduct,
 			featureQuantities,
-			resetCycleAnchor: startsAt,
+			entity,
+			resetCycleAnchor: startsAtSecondsPrecision,
 			freeTrial: null,
 			now: currentEpochMs,
 			billingVersion: BillingVersion.V2,
 		},
 		initOptions: {
-			startsAt,
-			endedAt: endsAt ?? undefined,
+			startsAt: startsAtSecondsPrecision,
+			endedAt: endsAtSecondsPrecision,
 			status: accessStartsAt === undefined ? CusProductStatus.Scheduled : undefined,
 			accessStartsAt,
 			externalId,
