+++ server/tests/integration/billing/create-schedule/create-schedule-phases.test.ts
@@ -1,30 +1,26 @@
 import { expect, test } from "bun:test";
 import {
-	type ApiCustomerV3,
-	type CreateScheduleParamsV0Input,
-	CheckoutAction,
-	CusProductStatus,
-	CustomerExpand,
-	customerEntitlements,
-	customerPrices,
-	customerProducts,
-	ms,
-	prices,
-	schedulePhases,
-	schedules,
+    type ApiCustomerV3,
+    type CreateScheduleParamsV0Input,
+    CheckoutAction,
+    CusProductStatus,
+    customerProducts,
+    ms,
+    schedulePhases,
+    schedules,
 } from "@autumn/shared";
 import {
-	confirmAutumnCheckout,
-	fetchAutumnCheckout,
+    confirmAutumnCheckout,
+    fetchAutumnCheckout,
 } from "@tests/integration/billing/utils/checkout/autumnCheckoutUtils";
 import { isAutumnCheckoutUrl } from "@tests/integration/billing/utils/isAutumnCheckoutUrl";
 import { completeInvoiceCheckoutV2 as completeInvoiceCheckout } from "@tests/utils/browserPool/completeInvoiceCheckoutV2";
 import { completeStripeCheckoutFormV2 as completeStripeCheckoutForm } from "@tests/utils/browserPool/completeStripeCheckoutFormV2";
+import { expectCustomerProducts } from "@tests/integration/billing/utils/expectCustomerProductCorrect";
 import { expectCustomerInvoiceCorrect } from "@tests/integration/billing/utils/expectCustomerInvoiceCorrect";
 import { TestFeature } from "@tests/setup/v2Features";
 import { expectAutumnError } from "@tests/utils/expectUtils/expectErrUtils";
 import { items } from "@tests/utils/fixtures/items";
-import { itemsV2 } from "@tests/utils/fixtures/itemsV2";
 import { products } from "@tests/utils/fixtures/products";
 import { advanceTestClock } from "@tests/utils/stripeUtils";
 import { timeout } from "@tests/utils/genUtils.js";
@@ -33,89 +29,15 @@ import chalk from "chalk";
 import { and, eq, inArray } from "drizzle-orm";
 import { CusService } from "@/internal/customers/CusService";
 import {
-	getFullCustomerSchedule,
-	hydrateCustomerWithSchedules,
+    getFullCustomerSchedule,
+    hydrateCustomerWithSchedules,
 } from "@/internal/customers/cusUtils/getFullCustomerSchedule";
 import { attachPaymentMethod } from "@/utils/scriptUtils/initCustomer";
-
-const getCustomerProductRows = async ({
-	ctx,
-	customerId,
-	productIds,
-}: {
-	ctx: Awaited<ReturnType<typeof initScenario>>["ctx"];
-	customerId: string;
-	productIds: string[];
-}) =>
-	await ctx.db
-		.select({
-			productId: customerProducts.product_id,
-			status: customerProducts.status,
-		})
-		.from(customerProducts)
-		.where(
-			and(
-				eq(customerProducts.customer_id, customerId),
-				inArray(customerProducts.product_id, productIds),
-			),
-		);
-
-const getCustomerProductPriceAmounts = async ({
-	ctx,
-	customerProductId,
-}: {
-	ctx: Awaited<ReturnType<typeof initScenario>>["ctx"];
-	customerProductId: string;
-}) =>
-	(
-		await ctx.db
-			.select({ config: prices.config })
-			.from(customerPrices)
-			.innerJoin(prices, eq(customerPrices.price_id, prices.id))
-			.where(eq(customerPrices.customer_product_id, customerProductId))
-	)
-		.map((row) =>
-			row.config && "amount" in row.config ? row.config.amount : undefined,
-		)
-		.filter((amount): amount is number => typeof amount === "number")
-		.sort((a, b) => a - b);
-
-const getCustomerProductEntitlementBalances = async ({
-	ctx,
-	customerProductId,
-}: {
-	ctx: Awaited<ReturnType<typeof initScenario>>["ctx"];
-	customerProductId: string;
-}) =>
-	await ctx.db
-		.select({
-			feature_id: customerEntitlements.feature_id,
-			balance: customerEntitlements.balance,
-		})
-		.from(customerEntitlements)
-		.where(eq(customerEntitlements.customer_product_id, customerProductId));
-
-const getRequiredScheduleId = (scheduleId: string | null) => {
-	if (!scheduleId) {
-		throw new Error("Expected create_schedule response to include schedule_id");
-	}
-
-	return scheduleId;
-};
-
-const getCheckoutId = (paymentUrl: string | null | undefined) => {
-	if (!paymentUrl) {
-		throw new Error("Expected create_schedule response to include payment_url");
-	}
-
-	const checkoutId = paymentUrl.split("/c/")[1];
-
-	if (!checkoutId) {
-		throw new Error(`Expected Autumn checkout URL, received: ${paymentUrl}`);
-	}
-
-	return checkoutId;
-};
+import {
+    getCheckoutId,
+    getCustomerProductRows,
+    getRequiredScheduleId,
+} from "./utils/createScheduleTestHelpers";
 
 test.concurrent(`${chalk.yellowBright("create-schedule: bills the first phase immediately and stores later phases as scheduled")}`, async () => {
 	const pro = products.pro({
@@ -328,416 +250,6 @@ test.concurrent(`${chalk.yellowBright("create-schedule: copies entity_id and rep
 	expect(newCustomerProducts).toHaveLength(1);
 	expect(newCustomerProducts[0]!.entity_id).toBe(entityId);
 	expect(newCustomerProducts[0]!.status).toBe(CusProductStatus.Active);
-});
-
-test.concurrent(`${chalk.yellowBright("create-schedule: preserves feature quantity options on created customer products")}`, async () => {
-	const prepaidMessages = products.base({
-		id: "prepaid",
-		items: [items.prepaidMessages()],
-	});
-
-	const { customerId, autumnV1, ctx } = await initScenario({
-		customerId: "create-schedule-options",
-		setup: [
-			s.customer({ paymentMethod: "success" }),
-			s.products({ list: [prepaidMessages] }),
-		],
-		actions: [],
-	});
-
-	const response = await autumnV1.billing.createSchedule({
-		customer_id: customerId,
-		phases: [
-			{
-				starts_at: Date.now(),
-				plans: [
-					{
-						plan_id: prepaidMessages.id,
-						feature_quantities: [
-							{
-								feature_id: TestFeature.Messages,
-								quantity: 400,
-							},
-						],
-					},
-				],
-			},
-		],
-	});
-
-	const insertedProducts = await ctx.db
-		.select({
-			options: customerProducts.options,
-		})
-		.from(customerProducts)
-		.where(
-			eq(customerProducts.id, response.phases[0]!.customer_product_ids[0]!),
-		);
-
-	expect(insertedProducts).toHaveLength(1);
-	expect(insertedProducts[0]!.options).toEqual([
-		expect.objectContaining({
-			feature_id: TestFeature.Messages,
-			quantity: 4,
-		}),
-	]);
-});
-
-test.concurrent(`${chalk.yellowBright("create-schedule: preserves customize.items on created customer products")}`, async () => {
-	const base = products.base({
-		id: "custom-base",
-		items: [items.monthlyMessages({ includedUsage: 100 })],
-	});
-
-	const { customerId, autumnV1, ctx } = await initScenario({
-		customerId: "create-schedule-customize",
-		setup: [
-			s.customer({ paymentMethod: "success" }),
-			s.products({ list: [base] }),
-		],
-		actions: [],
-	});
-
-	const response = await autumnV1.billing.createSchedule({
-		customer_id: customerId,
-		phases: [
-			{
-				starts_at: Date.now(),
-				plans: [
-					{
-						plan_id: base.id,
-						customize: {
-							items: [itemsV2.monthlyWords({ included: 250 })],
-						},
-					},
-				],
-			},
-		],
-	});
-
-	const insertedEntitlements = await ctx.db
-		.select({
-			feature_id: customerEntitlements.feature_id,
-			balance: customerEntitlements.balance,
-		})
-		.from(customerEntitlements)
-		.where(
-			eq(
-				customerEntitlements.customer_product_id,
-				response.phases[0]!.customer_product_ids[0]!,
-			),
-		);
-
-	expect(insertedEntitlements).toEqual([
-		{
-			feature_id: TestFeature.Words,
-			balance: 250,
-		},
-	]);
-});
-
-test.concurrent(`${chalk.yellowBright("create-schedule: preserves customize.items on future scheduled customer products")}`, async () => {
-	const base = products.base({
-		id: "custom-future-base",
-		items: [items.monthlyMessages({ includedUsage: 100 })],
-	});
-
-	const { customerId, autumnV1, ctx } = await initScenario({
-		customerId: "create-schedule-customize-future",
-		setup: [
-			s.customer({ paymentMethod: "success" }),
-			s.products({ list: [base] }),
-		],
-		actions: [],
-	});
-
-	const now = Date.now();
-	const response = await autumnV1.billing.createSchedule({
-		customer_id: customerId,
-		phases: [
-			{
-				starts_at: now,
-				plans: [{ plan_id: base.id }],
-			},
-			{
-				starts_at: now + ms.days(30),
-				plans: [
-					{
-						plan_id: base.id,
-						customize: {
-							items: [itemsV2.monthlyWords({ included: 250 })],
-						},
-					},
-				],
-			},
-		],
-	});
-
-	const futurePhase = response.phases[1];
-	expect(futurePhase).toBeDefined();
-
-	const futureCustomerProductId = futurePhase?.customer_product_ids[0];
-	expect(futureCustomerProductId).toBeTruthy();
-
-	if (!futureCustomerProductId) {
-		throw new Error(
-			"Expected a scheduled customer product for the future phase",
-		);
-	}
-
-	const insertedEntitlements = await ctx.db
-		.select({
-			feature_id: customerEntitlements.feature_id,
-			balance: customerEntitlements.balance,
-		})
-		.from(customerEntitlements)
-		.where(
-			eq(customerEntitlements.customer_product_id, futureCustomerProductId),
-		);
-
-	expect(insertedEntitlements).toEqual([
-		{
-			feature_id: TestFeature.Words,
-			balance: 250,
-		},
-	]);
-});
-
-test.concurrent(`${chalk.yellowBright("create-schedule: customized future phases keep custom prices and entitlements through activation")}`, async () => {
-	const base = products.base({
-		id: "create-schedule-customize-rollover",
-		items: [
-			items.monthlyMessages({ includedUsage: 100 }),
-			items.monthlyPrice(),
-		],
-	});
-
-	const { customerId, autumnV1, ctx, testClockId, advancedTo } =
-		await initScenario({
-			customerId: "create-schedule-customize-rollover",
-			setup: [
-				s.customer({ paymentMethod: "success" }),
-				s.products({ list: [base] }),
-			],
-			actions: [],
-		});
-
-	const now = advancedTo;
-	const response = await autumnV1.billing.createSchedule({
-		customer_id: customerId,
-		phases: [
-			{
-				starts_at: now,
-				plans: [{ plan_id: base.id }],
-			},
-			{
-				starts_at: now + ms.days(15),
-				plans: [
-					{
-						plan_id: base.id,
-						customize: {
-							price: itemsV2.monthlyPrice({ amount: 35 }),
-							items: [
-								itemsV2.monthlyWords({ included: 250 }),
-								itemsV2.dashboard(),
-							],
-						},
-					},
-				],
-			},
-		],
-	});
-
-	const futureCustomerProductId = response.phases[1]!.customer_product_ids[0]!;
-
-	expect(
-		await getCustomerProductPriceAmounts({
-			ctx,
-			customerProductId: futureCustomerProductId,
-		}),
-	).toEqual([35]);
-	expect(
-		await getCustomerProductEntitlementBalances({
-			ctx,
-			customerProductId: futureCustomerProductId,
-		}),
-	).toEqual(
-		expect.arrayContaining([
-			{ feature_id: TestFeature.Words, balance: 250 },
-			{ feature_id: TestFeature.Dashboard, balance: 0 },
-		]),
-	);
-
-	await advanceTestClock({
-		stripeCli: ctx.stripeCli,
-		testClockId: testClockId!,
-		advanceTo: now + ms.days(16),
-		waitForSeconds: 30,
-	});
-
-	const activatedProduct = await ctx.db.query.customerProducts.findFirst({
-		where: eq(customerProducts.id, futureCustomerProductId),
-	});
-
-	expect(activatedProduct?.status).toBe(CusProductStatus.Active);
-	expect(
-		await getCustomerProductPriceAmounts({
-			ctx,
-			customerProductId: futureCustomerProductId,
-		}),
-	).toEqual([35]);
-	expect(
-		await getCustomerProductEntitlementBalances({
-			ctx,
-			customerProductId: futureCustomerProductId,
-		}),
-	).toEqual(
-		expect.arrayContaining([
-			{ feature_id: TestFeature.Words, balance: 250 },
-			{ feature_id: TestFeature.Dashboard, balance: 0 },
-		]),
-	);
-});
-
-test.concurrent(`${chalk.yellowBright("create-schedule: updating a future customized phase replaces its custom prices and quantities before activation")}`, async () => {
-	const base = products.base({
-		id: "create-schedule-custom-update",
-		items: [
-			items.monthlyMessages({ includedUsage: 100 }),
-			items.monthlyPrice(),
-		],
-	});
-
-	const { customerId, autumnV1, ctx, testClockId, advancedTo } =
-		await initScenario({
-			customerId: "create-schedule-custom-update",
-			setup: [
-				s.customer({ paymentMethod: "success" }),
-				s.products({ list: [base] }),
-			],
-			actions: [],
-		});
-
-	const now = advancedTo;
-	const initialResponse = await autumnV1.billing.createSchedule({
-		customer_id: customerId,
-		phases: [
-			{
-				starts_at: now,
-				plans: [{ plan_id: base.id }],
-			},
-			{
-				starts_at: now + ms.days(15),
-				plans: [
-					{
-						plan_id: base.id,
-						customize: {
-							price: itemsV2.monthlyPrice({ amount: 35 }),
-							items: [
-								itemsV2.prepaidMessages({ amount: 10, billingUnits: 100 }),
-							],
-						},
-						feature_quantities: [
-							{
-								feature_id: TestFeature.Messages,
-								quantity: 200,
-							},
-						],
-					},
-				],
-			},
-		],
-	});
-
-	const initialFutureCustomerProductId =
-		initialResponse.phases[1]!.customer_product_ids[0]!;
-
-	const updatedResponse = await autumnV1.billing.createSchedule({
-		customer_id: customerId,
-		phases: [
-			{
-				starts_at: now,
-				plans: [{ plan_id: base.id }],
-			},
-			{
-				starts_at: now + ms.days(15),
-				plans: [
-					{
-						plan_id: base.id,
-						customize: {
-							price: itemsV2.monthlyPrice({ amount: 55 }),
-							items: [
-								itemsV2.prepaidMessages({ amount: 20, billingUnits: 100 }),
-							],
-						},
-						feature_quantities: [
-							{
-								feature_id: TestFeature.Messages,
-								quantity: 500,
-							},
-						],
-					},
-				],
-			},
-		],
-	});
-
-	const updatedFutureCustomerProductId =
-		updatedResponse.phases[1]!.customer_product_ids[0]!;
-
-	expect(updatedFutureCustomerProductId).not.toBe(
-		initialFutureCustomerProductId,
-	);
-	expect(
-		await ctx.db.query.customerProducts.findFirst({
-			where: eq(customerProducts.id, initialFutureCustomerProductId),
-		}),
-	).toBeUndefined();
-
-	const updatedFutureCustomerProduct =
-		await ctx.db.query.customerProducts.findFirst({
-			where: eq(customerProducts.id, updatedFutureCustomerProductId),
-		});
-
-	expect(updatedFutureCustomerProduct?.status).toBe(CusProductStatus.Scheduled);
-	expect(updatedFutureCustomerProduct?.options).toEqual([
-		expect.objectContaining({
-			feature_id: TestFeature.Messages,
-			quantity: 5,
-		}),
-	]);
-	expect(
-		await getCustomerProductPriceAmounts({
-			ctx,
-			customerProductId: updatedFutureCustomerProductId,
-		}),
-	).toEqual([55]);
-
-	await advanceTestClock({
-		stripeCli: ctx.stripeCli,
-		testClockId: testClockId!,
-		advanceTo: now + ms.days(16),
-		waitForSeconds: 30,
-	});
-
-	const activatedFutureCustomerProduct =
-		await ctx.db.query.customerProducts.findFirst({
-			where: eq(customerProducts.id, updatedFutureCustomerProductId),
-		});
-
-	expect(activatedFutureCustomerProduct?.status).toBe(CusProductStatus.Active);
-	expect(activatedFutureCustomerProduct?.options).toEqual([
-		expect.objectContaining({
-			feature_id: TestFeature.Messages,
-			quantity: 5,
-		}),
-	]);
-	expect(
-		await getCustomerProductPriceAmounts({
-			ctx,
-			customerProductId: updatedFutureCustomerProductId,
-		}),
-	).toEqual([55]);
 });
 
 test.concurrent(`${chalk.yellowBright("create-schedule: persists the new schedule and returns required_action when immediate billing is deferred")}`, async () => {
@@ -1237,42 +749,12 @@ test.concurrent(`${chalk.yellowBright("create-schedule: plans omitted from the n
 		waitForSeconds: 30,
 	});
 
-	const productRows = await getCustomerProductRows({
-		ctx,
-		customerId,
-		productIds: [nowBase.id, nowAddon.id, nextBase.id, nextAddon.id],
-	});
-
-	expect(
-		productRows
-			.filter((productRow) => productRow.status === CusProductStatus.Active)
-			.sort((a, b) => a.productId!.localeCompare(b.productId!)),
-	).toEqual(
-		[
-			{ productId: nextAddon.id, status: CusProductStatus.Active },
-			{ productId: nextBase.id, status: CusProductStatus.Active },
-		].sort((a, b) => a.productId.localeCompare(b.productId)),
-	);
-	expect(
-		productRows.filter(
-			(productRow) => productRow.status === CusProductStatus.Scheduled,
-		),
-	).toHaveLength(0);
-	expect(
-		productRows
-			.filter((productRow) => productRow.status === CusProductStatus.Expired)
-			.sort((a, b) => a.productId!.localeCompare(b.productId!)),
-	).toEqual(
-		[
-			{ productId: nowAddon.id, status: CusProductStatus.Expired },
-			{ productId: nowBase.id, status: CusProductStatus.Expired },
-		].sort((a, b) => a.productId.localeCompare(b.productId)),
-	);
-
 	const customer = await autumnV1.customers.get<ApiCustomerV3>(customerId);
-	expect(customer.products?.map((product) => product.id).sort()).toEqual(
-		[nextAddon.id, nextBase.id].sort(),
-	);
+	await expectCustomerProducts({
+		customer,
+		active: [nextAddon.id, nextBase.id],
+		notPresent: [nowAddon.id, nowBase.id],
+	});
 });
 
 test.concurrent(`${chalk.yellowBright("create-schedule: rejects updating a schedule after earlier phases started when past phases are resubmitted")}`, async () => {
@@ -1308,7 +790,7 @@ test.concurrent(`${chalk.yellowBright("create-schedule: rejects updating a sched
 		});
 
 	const now = advancedTo;
-	const initialResponse = await autumnV1.billing.createSchedule({
+	await autumnV1.billing.createSchedule({
 		customer_id: customerId,
 		phases: [
 			{
@@ -1979,9 +1461,9 @@ test.concurrent(`${chalk.yellowBright("create-schedule: hydrates schedules on th
 	expect(hydratedCustomer.schedule?.customer_id).toBe(customerId);
 	expect(hydratedCustomer.schedule?.phases).toHaveLength(2);
 	expect(hydratedCustomer.schedule?.phases[0]?.starts_at).toBe(now);
-expect(hydratedCustomer.schedule?.phases[1]?.starts_at).toBe(
-	now + ms.days(30),
-);
+	expect(hydratedCustomer.schedule?.phases[1]?.starts_at).toBe(
+		now + ms.days(30),
+	);
 });
 
 test.concurrent(`${chalk.yellowBright("create-schedule: adding a future phase to an existing single-phase schedule persists both phases")}`, async () => {
@@ -2085,150 +1567,6 @@ test.concurrent(`${chalk.yellowBright("create-schedule: adding a future phase to
 	expect(futureProducts[0]!.product_id).toBe(premium.id);
 });
 
-test.concurrent(`${chalk.yellowBright("create-schedule: updating a schedule with customized future phase persists both phases and custom items")}`, async () => {
-	const base = products.base({
-		id: "base",
-		items: [
-			items.monthlyMessages({ includedUsage: 100 }),
-			items.monthlyPrice(),
-		],
-	});
-
-	const { customerId, autumnV1, ctx } = await initScenario({
-		customerId: "create-schedule-update-with-customize",
-		setup: [
-			s.customer({ paymentMethod: "success" }),
-			s.products({ list: [base] }),
-		],
-		actions: [],
-	});
-
-	const now = Date.now();
-	const initialResponse = await autumnV1.billing.createSchedule({
-		customer_id: customerId,
-		phases: [
-			{
-				starts_at: now,
-				plans: [{ plan_id: base.id }],
-			},
-		],
-	});
-
-	expect(initialResponse.phases).toHaveLength(1);
-
-	const updatedResponse = await autumnV1.billing.createSchedule({
-		customer_id: customerId,
-		phases: [
-			{
-				starts_at: now,
-				plans: [{ plan_id: base.id }],
-			},
-			{
-				starts_at: now + ms.days(30),
-				plans: [
-					{
-						plan_id: base.id,
-						customize: {
-							price: itemsV2.monthlyPrice({ amount: 50 }),
-							items: [itemsV2.monthlyWords({ included: 200 })],
-						},
-					},
-				],
-			},
-		],
-	});
-
-	expect(updatedResponse.phases).toHaveLength(2);
-
-	const updatedDbPhases = await ctx.db
-		.select()
-		.from(schedulePhases)
-		.where(
-			eq(
-				schedulePhases.schedule_id,
-				getRequiredScheduleId(updatedResponse.schedule_id),
-			),
-		);
-	expect(updatedDbPhases).toHaveLength(2);
-
-	const futureCustomerProductId =
-		updatedResponse.phases[1]!.customer_product_ids[0]!;
-
-	expect(
-		await getCustomerProductPriceAmounts({
-			ctx,
-			customerProductId: futureCustomerProductId,
-		}),
-	).toEqual([50]);
-
-	expect(
-		await getCustomerProductEntitlementBalances({
-			ctx,
-			customerProductId: futureCustomerProductId,
-		}),
-	).toEqual(
-		expect.arrayContaining([{ feature_id: TestFeature.Words, balance: 200 }]),
-	);
-});
-
-test.concurrent(`${chalk.yellowBright("create-schedule: customize with boolean feature persists the boolean entitlement")}`, async () => {
-	const base = products.base({
-		id: "bool-base",
-		items: [
-			items.monthlyMessages({ includedUsage: 100 }),
-			items.monthlyPrice(),
-		],
-	});
-
-	const { customerId, autumnV1, ctx } = await initScenario({
-		customerId: "create-schedule-customize-boolean",
-		setup: [
-			s.customer({ paymentMethod: "success" }),
-			s.products({ list: [base] }),
-		],
-		actions: [],
-	});
-
-	const response = await autumnV1.billing.createSchedule({
-		customer_id: customerId,
-		phases: [
-			{
-				starts_at: Date.now(),
-				plans: [
-					{
-						plan_id: base.id,
-						customize: {
-							items: [
-								itemsV2.monthlyMessages({ included: 100 }),
-								itemsV2.dashboard(),
-							],
-						},
-					},
-				],
-			},
-		],
-	});
-
-	const customerProductId = response.phases[0]!.customer_product_ids[0]!;
-
-	const entitlementBalances = await getCustomerProductEntitlementBalances({
-		ctx,
-		customerProductId,
-	});
-
-	expect(entitlementBalances).toEqual(
-		expect.arrayContaining([
-			{ feature_id: TestFeature.Messages, balance: 100 },
-			{ feature_id: TestFeature.Dashboard, balance: 0 },
-		]),
-	);
-
-	const customerProduct = await ctx.db.query.customerProducts.findFirst({
-		where: eq(customerProducts.id, customerProductId),
-	});
-	expect(customerProduct?.is_custom).toBe(true);
-});
-
 test.concurrent(`${chalk.yellowBright("create-schedule: customer-level and entity-level schedules coexist independently")}`, async () => {
 	const pro = products.pro({
 		id: "pro",
@@ -2259,6 +1597,10 @@ test.concurrent(`${chalk.yellowBright("create-schedule: customer-level and entit
 				starts_at: now,
 				plans: [{ plan_id: pro.id }],
 			},
+			{
+				starts_at: now + ms.days(30),
+				plans: [{ plan_id: pro.id }],
+			},
 		],
 	});
 
@@ -2268,6 +1610,10 @@ test.concurrent(`${chalk.yellowBright("create-schedule: customer-level and entit
 		phases: [
 			{
 				starts_at: now,
+				plans: [{ plan_id: addon.id }],
+			},
+			{
+				starts_at: now + ms.days(30),
 				plans: [{ plan_id: addon.id }],
 			},
 		],
