--- vite/src/views/customers2/utils/effectiveCustomerProductStatus.ts
@@ -1,19 +1,0 @@
-import {
-	CusProductStatus,
-	type FullCusProduct,
-	hasCustomerProductEnded,
-} from "@autumn/shared";
-
-export function withEffectiveCustomerProductStatus({
-	customerProduct,
-	nowMs,
-}: {
-	customerProduct: FullCusProduct;
-	nowMs?: number;
-}): FullCusProduct {
-	if (!hasCustomerProductEnded(customerProduct, { nowMs })) {
-		return customerProduct;
-	}
-
-	return { ...customerProduct, status: CusProductStatus.Expired };
-}
